requirements.txt
ipykernel==7.0.1
pillow==12.0.0 # 图像生成模块使用
addict==2.4.0
transformers==4.46.3
tokenizers==0.20.3
tiktoken=0.12.0
protobuf==6.33.0